split graph tests into 2 specs
