+++ packages/plot-graph/import-map.json
@@ -1,3 +1,76 @@
 {
-    "modules": {}
+    "modules": {
+        "@hmh/component-base": "/node_modules/@hmh/component-base/dist/index.js",
+        "@polymer/app-layout/app-drawer-layout/app-drawer-layout.js": "/node_modules/@polymer/app-layout/app-drawer-layout/app-drawer-layout.js",
+        "@polymer/app-layout/app-drawer/app-drawer.js": "/node_modules/@polymer/app-layout/app-drawer/app-drawer.js",
+        "@polymer/app-layout/app-header-layout/app-header-layout.js": "/node_modules/@polymer/app-layout/app-header-layout/app-header-layout.js",
+        "@polymer/app-layout/app-header/app-header.js": "/node_modules/@polymer/app-layout/app-header/app-header.js",
+        "@polymer/app-layout/app-toolbar/app-toolbar.js": "/node_modules/@polymer/app-layout/app-toolbar/app-toolbar.js",
+        "@polymer/font-roboto/roboto.js": "/node_modules/@polymer/font-roboto/roboto.js",
+        "@polymer/iron-a11y-announcer/iron-a11y-announcer.js": "/node_modules/@polymer/iron-a11y-announcer/iron-a11y-announcer.js",
+        "@polymer/iron-a11y-keys-behavior/iron-a11y-keys-behavior.js": "/node_modules/@polymer/iron-a11y-keys-behavior/iron-a11y-keys-behavior.js",
+        "@polymer/iron-ajax/iron-ajax.js": "/node_modules/@polymer/iron-ajax/iron-ajax.js",
+        "@polymer/iron-behaviors/iron-button-state.js": "/node_modules/@polymer/iron-behaviors/iron-button-state.js",
+        "@polymer/iron-behaviors/iron-control-state.js": "/node_modules/@polymer/iron-behaviors/iron-control-state.js",
+        "@polymer/iron-component-page": "/node_modules/@polymer/iron-component-page/iron-component-page.js",
+        "@polymer/iron-doc-viewer/default-theme.js": "/node_modules/@polymer/iron-doc-viewer/default-theme.js",
+        "@polymer/iron-doc-viewer/iron-doc-nav.js": "/node_modules/@polymer/iron-doc-viewer/iron-doc-nav.js",
+        "@polymer/iron-doc-viewer/iron-doc-viewer.js": "/node_modules/@polymer/iron-doc-viewer/iron-doc-viewer.js",
+        "@polymer/iron-fit-behavior/iron-fit-behavior.js": "/node_modules/@polymer/iron-fit-behavior/iron-fit-behavior.js",
+        "@polymer/iron-flex-layout/iron-flex-layout.js": "/node_modules/@polymer/iron-flex-layout/iron-flex-layout.js",
+        "@polymer/iron-icon/iron-icon.js": "/node_modules/@polymer/iron-icon/iron-icon.js",
+        "@polymer/iron-icons/iron-icons.js": "/node_modules/@polymer/iron-icons/iron-icons.js",
+        "@polymer/iron-iconset-svg/iron-iconset-svg.js": "/node_modules/@polymer/iron-iconset-svg/iron-iconset-svg.js",
+        "@polymer/iron-location/iron-location.js": "/node_modules/@polymer/iron-location/iron-location.js",
+        "@polymer/iron-media-query/iron-media-query.js": "/node_modules/@polymer/iron-media-query/iron-media-query.js",
+        "@polymer/iron-meta/iron-meta.js": "/node_modules/@polymer/iron-meta/iron-meta.js",
+        "@polymer/iron-overlay-behavior/iron-overlay-behavior.js": "/node_modules/@polymer/iron-overlay-behavior/iron-overlay-behavior.js",
+        "@polymer/iron-resizable-behavior/iron-resizable-behavior.js": "/node_modules/@polymer/iron-resizable-behavior/iron-resizable-behavior.js",
+        "@polymer/iron-scroll-target-behavior/iron-scroll-target-behavior.js": "/node_modules/@polymer/iron-scroll-target-behavior/iron-scroll-target-behavior.js",
+        "@polymer/lit-element": "/node_modules/@hmh/component-base/node_modules/@polymer/lit-element/lit-element.js",
+        "@polymer/marked-element/marked-element.js": "/node_modules/@polymer/marked-element/marked-element.js",
+        "@polymer/paper-behaviors/paper-inky-focus-behavior.js": "/node_modules/@polymer/paper-behaviors/paper-inky-focus-behavior.js",
+        "@polymer/paper-icon-button/paper-icon-button.js": "/node_modules/@polymer/paper-icon-button/paper-icon-button.js",
+        "@polymer/paper-ripple/paper-ripple.js": "/node_modules/@polymer/paper-ripple/paper-ripple.js",
+        "@polymer/paper-styles/color.js": "/node_modules/@polymer/paper-styles/color.js",
+        "@polymer/paper-styles/default-theme.js": "/node_modules/@polymer/paper-styles/default-theme.js",
+        "@polymer/paper-styles/typography.js": "/node_modules/@polymer/paper-styles/typography.js",
+        "@polymer/paper-toast/paper-toast.js": "/node_modules/@polymer/paper-toast/paper-toast.js",
+        "@polymer/polymer/lib/elements/dom-module.js": "/node_modules/@polymer/polymer/lib/elements/dom-module.js",
+        "@polymer/polymer/lib/legacy/polymer-fn.js": "/node_modules/@polymer/polymer/lib/legacy/polymer-fn.js",
+        "@polymer/polymer/lib/legacy/polymer.dom.js": "/node_modules/@polymer/polymer/lib/legacy/polymer.dom.js",
+        "@polymer/polymer/lib/utils/async.js": "/node_modules/@polymer/polymer/lib/utils/async.js",
+        "@polymer/polymer/lib/utils/debounce.js": "/node_modules/@polymer/polymer/lib/utils/debounce.js",
+        "@polymer/polymer/lib/utils/flush.js": "/node_modules/@polymer/polymer/lib/utils/flush.js",
+        "@polymer/polymer/lib/utils/gestures.js": "/node_modules/@polymer/polymer/lib/utils/gestures.js",
+        "@polymer/polymer/lib/utils/html-tag.js": "/node_modules/@polymer/polymer/lib/utils/html-tag.js",
+        "@polymer/polymer/lib/utils/render-status.js": "/node_modules/@polymer/polymer/lib/utils/render-status.js",
+        "@polymer/polymer/lib/utils/settings.js": "/node_modules/@polymer/polymer/lib/utils/settings.js",
+        "@polymer/polymer/polymer-legacy.js": "/node_modules/@polymer/polymer/polymer-legacy.js",
+        "@polymer/prism-element/prism-highlighter.js": "/node_modules/@polymer/prism-element/prism-highlighter.js",
+        "@polymer/prism-element/prism-theme-default.js": "/node_modules/@polymer/prism-element/prism-theme-default.js",
+        "@webcomponents/shadycss/entrypoints/apply-shim.js": "/node_modules/@webcomponents/shadycss/entrypoints/apply-shim.js",
+        "@webcomponents/shadycss/entrypoints/custom-style-interface.js": "/node_modules/@webcomponents/shadycss/entrypoints/custom-style-interface.js",
+        "d3-array": "/node_modules/d3-array/src/index.js",
+        "d3-axis": "/node_modules/d3-axis/src/index.js",
+        "d3-collection": "/node_modules/d3-collection/src/index.js",
+        "d3-color": "/node_modules/d3-color/src/index.js",
+        "d3-format": "/node_modules/d3-format/src/index.js",
+        "d3-interpolate": "/node_modules/d3-interpolate/src/index.js",
+        "d3-path": "/node_modules/d3-path/src/index.js",
+        "d3-scale": "/node_modules/d3-scale/src/index.js",
+        "d3-selection": "/node_modules/d3-selection/src/index.js",
+        "d3-shape": "/node_modules/d3-shape/src/index.js",
+        "d3-time": "/node_modules/d3-time/src/index.js",
+        "d3-time-format": "/node_modules/d3-time-format/src/index.js",
+        "lit-html": "/node_modules/@hmh/component-base/node_modules/lit-html/lit-html.js",
+        "lit-html/directives/if-defined.js": "/node_modules/@hmh/component-base/node_modules/lit-html/directives/if-defined.js",
+        "lit-html/directives/repeat.js": "/node_modules/@hmh/component-base/node_modules/lit-html/directives/repeat.js",
+        "lit-html/directives/unsafe-html.js": "/node_modules/@hmh/component-base/node_modules/lit-html/directives/unsafe-html.js",
+        "lit-html/lib/shady-render": "/node_modules/@hmh/component-base/node_modules/lit-html/lib/shady-render.js",
+        "lit-html/lit-html": "/node_modules/@hmh/component-base/node_modules/lit-html/lit-html.js",
+        "marked/lib/marked.js": "/node_modules/marked/lib/marked.js",
+        "prismjs/prism.js": "/node_modules/prismjs/prism.js",
+        "vis/dist/vis.min.css": "/node_modules/vis/dist/vis.min.css"
+    }
 }
