+++ packages/plot-graph/src/unit/plot-graph-3d.spec.ts
@@ -3,14 +3,14 @@ import { checkComponentDOM } from './test-helpers.js';
 const tagName: string = 'plot-graph-3d';
 // const expect: any = chai.expect;
 
-describe(`<${tagName}>`, (): void => {
-    it('should render default state (3d)', async (): Promise<void> => {
-        withSnippet('default');
-        const el: PlotGraph3D = document.querySelector('plot-graph-3d') as any;
-        // @ts-ignore
-        await el.updateComplete;
-        checkComponentDOM(el);
+export default () => {
+    describe(`<${tagName}>`, (): void => {
+        it.only('should render default state (3d)', async (): Promise<void> => {
+            withSnippet('default-3d');
+            const el: PlotGraph3D = document.querySelector(tagName) as any;
+            // @ts-ignore
+            await el.updateComplete;
+            checkComponentDOM(el);
+        });
     });
-});
-
-mocha.run();
+}
